Extract AuthorsCombobox component
